move files to appropriate folder
